backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
psycopg2-binary==2.9.9
setuptools==74.1.1
sqlparse==0.5.1
wheel==0.37.1
djangorestframework~=3.15.2